Got commit-msg hook working
Pretty simple example of a git commit-msg hook that will validate all
youtrack issues inside of a commit message. Also shows an example of how
to add validation of custom fields by requiring a youtrack issue type of
feature to be approved for work. Also checks subtasks of a feature and
verifys the feature itself is verified.